move some things
